// src/api/models/resume-im.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online Resume Builder API
 * An API for Online Resume Builder
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface ResumeIM
 */
export interface ResumeIM {
    /**
     * 
     * @type {string}
     * @memberof ResumeIM
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof ResumeIM
     */
    templateId: string;
}